+++ packages/wonder-blocks-progress-spinner/src/components/circular-spinner.js
@@ -58,6 +58,14 @@ type DefaultProps = {|
 /**
  * A circular progress spinner. Used for indicating loading progress. Should
  * be used by default in most places where a loading indicator is needed.
+ *
+ * ### Usage
+ *
+ * ```js
+ * import {CircularSpinner} from "@khanacademy/wonder-blocks-progress-spinner";
+ *
+ * <CircularSpinner />
+ * ```
  */
 export default class CircularSpinner extends React.Component<Props> {
     static defaultProps: DefaultProps = {