+++ packages/wonder-blocks-popover/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-popover
 
+## 1.1.17
+
+### Patch Changes
+
+-   Updated dependencies [6ee20af9]
+    -   @khanacademy/wonder-blocks-core@4.4.0
+    -   @khanacademy/wonder-blocks-icon@1.2.30
+    -   @khanacademy/wonder-blocks-icon-button@3.4.11
+    -   @khanacademy/wonder-blocks-layout@1.4.11
+    -   @khanacademy/wonder-blocks-modal@2.3.7
+    -   @khanacademy/wonder-blocks-tooltip@1.3.17
+    -   @khanacademy/wonder-blocks-typography@1.1.33
+
+## 1.1.16
+
+### Patch Changes
+
+-   Updated dependencies [2546b126]
+-   Updated dependencies [e9364406]
+    -   @khanacademy/wonder-blocks-modal@2.3.6
+    -   @khanacademy/wonder-blocks-tooltip@1.3.16
+
 ## 1.1.15
 
 ### Patch Changes